Renamed outstream factory file, added to_string.
