+++ src/instrumentation/outstream/tests/outstream_factory_integration_test.cpp
@@ -1,4 +1,4 @@
-#include "instrumentation/outstream/factory.h"
+#include "instrumentation/outstream/factory.hpp"
 
 #include <memory>
 #include <sstream>
